Fix numerous errors across the website and core libraries
Addresses 33 errors by updating class definitions, method implementations, type casting, and resolving reference issues across multiple .cs files, including stubs for configuration, moderation, and server libraries.

+++ RobloxWebSite/ModelFactories/Moderation/PunishmentModelFactory.cs
@@ -23,7 +23,7 @@ namespace Roblox.Website.ModelFactories.Moderation
             {
                 ID = punishment.ID,
                 Action = punishment.PunishmentType.Value,
-                ModeratorID = punishment.ModeratorID,
+                ModeratorID = (int)punishment.ModeratorID,
                 ModeratorName = moderator.Name,
                 InternalNote = punishment.Comment,
                 ModeratorMessage = punishment.ModeratorMessage,
